fix(frontend): update serve command syntax and move env validation to build-time
