+++ frontend/Dockerfile
@@ -2,9 +2,8 @@ FROM node:24-alpine
 WORKDIR /app
 COPY package*.json ./
 RUN npm ci
+# TODO: Optimize this dockerfile by using a multi-stage build
+# TODO: Mount source code for development
 COPY . .
-# Run the expo server
-# TODO: For prod, it is better to export the app and serve the static files. 
-# However, the expo .env variables are not loaded in the build phase.
-# For now, we will run the expo server in development mode.
+# Default to development server. This can be overridden for production builds.
 CMD ["npx", "expo", "start", "--port", "3000", "--web"]
