+++ frontend/package.json
@@ -51,6 +51,7 @@
     "react-native-screens": "~4.11.1",
     "react-native-web": "^0.20.0",
     "react-native-webview": "13.13.5",
+    "serve": "^14.2.4",
     "typescript": "~5.8.3",
     "zustand": "^5.0.5"
   },
